+++ check/src/board_code.cpp
@@ -26,8 +26,8 @@ namespace wisdom
 
         if (is_castling_move (move))
         {
-            int8_t src_col, dst_col;
-            int8_t row;
+            int src_col, dst_col;
+            int row;
 
             if (is_castling_move_on_king_side (move))
             {
@@ -81,8 +81,8 @@ namespace wisdom
 
         if (is_castling_move (move))
         {
-            int8_t src_col, dst_col;
-            int8_t row;
+            int src_col, dst_col;
+            int row;
 
             if (is_castling_move_on_king_side (move))
             {
