+++ check/src/check.hpp
@@ -12,7 +12,7 @@ namespace wisdom
 
     // check if the the king indicated by the WHO argument is in trouble
     // in this position
-    bool is_king_threatened (Board &board, Color who,int8_t row, int8_t col);
+    bool is_king_threatened (Board &board, Color who, int row, int col);
 
     static inline bool is_king_threatened (Board &board, Color who, Coord pos)
     {
